Rename backends-data/Clovis -> backends-data/Motr
